+++ src/index.js
@@ -3,7 +3,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import {createStore} from "redux";
 import {Provider} from "react-redux";
-import {reducer} from "./reducer.js";
+import {reducer} from "./redux/reducer";
 
 
 const store = createStore(reducer,
