--- src/utils.js
@@ -1,3 +1,0 @@
-export const extend = (a, b) => {
-  return Object.assign({}, a, b);
-};
